requirements.txt
aiodns==1.1.1
aiohttp==2.3.6
aiohttp-cors==0.5.3
aiohttp-debugtoolbar==0.4.1
aiohttp-devtools==0.6.4
aiohttp-jinja2==0.14.0
aiohttp-utils==3.0.0
aiojobs==0.0.3
async-timeout==2.0.0
cchardet==2.1.1
chardet==3.0.4
click==6.7
gunicorn==19.7.1
isort==4.2.15
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.14.0
motor==1.1
multidict==3.3.2
pycares==2.3.0
PyJWT==1.5.3
pymongo==3.6.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
six==1.11.0
typing==3.6.2
umongo==0.15.0
watchgod==0.0.3
Werkzeug==0.12.2
yarl==0.15.0
